* 'develop' of https://github.com/ONLYOFFICE/core:
  RtfFormat - add office digital signatures
  bug with some jpeg
  x2t version up 2.0.3.438
  RtfFormat - fix croping image, ...
  RtfFormat - extended drawing
  fix bug #34016
  fix bug #34018
  fix bug 34011
  fix bug #33752 after refactoring
  fix convert oox group shape to vml group shape
  RtfFormatWriter - fix table cell width (auto)
  OdfFormat - Reader/Writer - background page (image, pattern, gradient)
  fix bug #33918

+++ ASCOfficeDocxFile2/DocWrapper/FontProcessor.cpp
@@ -58,7 +58,7 @@ namespace DocWrapper {
 	{
 		//подбор перенесен в js
 		return;
-		for (int i = 0; i < fontTable->m_arrFonts.size(); ++i)
+		for (size_t i = 0; i < fontTable->m_arrFonts.size(); ++i)
 			addToFontMap(*fontTable->m_arrFonts[i]);
 	}
