+++ ASCOfficeRtfFile/RtfFormatLib/source/Reader/OOXNumberingMapReader.h
@@ -54,7 +54,7 @@ public:
 		if (m_ooxNum->m_oAbstractNumId.IsInit() && m_ooxNum->m_oAbstractNumId->m_oVal.IsInit())
 			oNewOverride.m_nListID = m_ooxNum->m_oAbstractNumId->m_oVal->GetValue();
 
-		for (long i=0 ;i < m_ooxNum->m_arrLvlOverride.size(); i++)
+		for (size_t i = 0 ; i < m_ooxNum->m_arrLvlOverride.size(); i++)
 		{
 			if (m_ooxNum->m_arrLvlOverride[i] == NULL) continue;
