+++ Common/DocxFormat/Source/DocxFormat/Styles.h
@@ -819,7 +819,7 @@ namespace OOX
 						{
 							if (oStyle->m_oName.IsInit())
 							{
-								m_arrStyleNamesMap[oStyle->m_oName->ToString()] = m_arrStyle.size();
+								m_arrStyleNamesMap[oStyle->m_oName->ToString()] = (int)m_arrStyle.size();
 							}
 							m_arrStyle.push_back( oStyle );
 						}
